+++ app/main.py
@@ -2,11 +2,11 @@ from fastapi import Depends, FastAPI, HTTPException
 from fastapi.middleware.cors import CORSMiddleware
 from sqlalchemy.orm import Session
 from typing import Annotated
-from db import crud
-from db import models
-from db import schemas
-from db.database import SessionLocal, engine
-import auth
+from app.db import crud
+from app.db import models
+from app.db import schemas
+from app.db.database import SessionLocal, engine
+import app.routes.auth as auth
 from logs.logging_config import setup_logging
 
 models.Base.metadata.create_all(bind=engine)
